+++ resources/ts/Components/Forms/CountryDropdown.tsx
@@ -1,6 +1,6 @@
 import { ChangeEventHandler, useEffect, useState } from "react";
 import { allCountries, CountryData } from "country-region-data";
-import SelectInput from "@/Components/Forms/SelectInput";
+import SelectInput from "@/Elements/Forms/SelectInput";
 
 export default function CountryDropdown({ value, handleChange }: { value: string, handleChange: ChangeEventHandler<HTMLElement> }) {
     const [countryList, setCountryList] = useState<CountryData[]>([]);
