+++ resources/ts/Elements/Forms/SelectInput.tsx
@@ -1,28 +1,21 @@
 import { ChangeEventHandler } from "react";
 
 interface Select {
-    name: string;
-    id: string;
     value?: string;
     className?: string;
-    required?: boolean | undefined;
     handleChange: ChangeEventHandler<HTMLSelectElement>
     children: React.ReactNode,
     defaultValue?: string;
 }
 
 
-export default function SelectInput({ name, id, value, className = "", required, handleChange, children, defaultValue }: Select) {
+export default function SelectInput({ value, className = "", handleChange, children, defaultValue, ...props }: Select & React.SelectHTMLAttributes<HTMLSelectElement>) {
     return (
         <select
-            name={name}
-            id={id}
+            {...props}
             defaultValue={!value ? defaultValue : undefined}
-            value={value}
             className={`border-gray-400 bg-clip-padding focus:border-indigo-500 focus:ring-indigo-500 rounded-md shadow-sm transition ease-in-out text-gray-700 focus-within:text-inherit ${className}`}
-            onChange={(e) => handleChange(e)}
-            autoComplete="on"
-            required={required}>
+            onChange={(e) => handleChange(e)}>
             {children}
         </select>
     )
