+++ resources/ts/Pages/Assembly/Edit.tsx
@@ -1,9 +1,9 @@
 import ButtonLink from "@/Elements/Buttons/ButtonLink";
 import PrimaryButton from "@/Elements/Buttons/PrimaryButton";
 import SecondaryButton from "@/Elements/Buttons/SecondaryButton";
-import FileInput from "@/Components/Forms/FileInput";
-import InputLabel from "@/Components/Forms/InputLabel";
-import TextInput from "@/Components/Forms/TextInput";
+import FileInput from "@/Elements/Forms/FileInput";
+import InputLabel from "@/Elements/Forms/InputLabel";
+import TextInput from "@/Elements/Forms/TextInput";
 import ToastBanner from "@/Components/Forms/ToastBanner";
 import ContentWrapper from "@/Layouts/ContentWrapper";
 import WrapperLayout from "@/Layouts/WrapperLayout";
