+++ resources/ts/Pages/Events/Step/Past/Create.tsx
@@ -1,13 +1,13 @@
 import ButtonLink from "@/Elements/Buttons/ButtonLink";
 import PrimaryButton from "@/Elements/Buttons/PrimaryButton";
 import SecondaryButton from "@/Elements/Buttons/SecondaryButton";
-import FileInput from "@/Components/Forms/FileInput";
-import InputLabel from "@/Components/Forms/InputLabel";
-import InputLabel2 from "@/Components/Forms/InputLabel2";
-import Legend from "@/Components/Forms/Legend";
-import RadioInput from "@/Components/Forms/RadioInput";
-import TextAreaInput from "@/Components/Forms/TextAreaInput";
-import TextInput from "@/Components/Forms/TextInput";
+import FileInput from "@/Elements/Forms/FileInput";
+import InputLabel from "@/Elements/Forms/InputLabel";
+import InputLabel2 from "@/Elements/Forms/InputLabel2";
+import Legend from "@/Elements/Forms/Legend";
+import RadioInput from "@/Elements/Forms/RadioInput";
+import TextAreaInput from "@/Elements/Forms/TextAreaInput";
+import TextInput from "@/Elements/Forms/TextInput";
 import ToastBanner from "@/Components/Forms/ToastBanner";
 import VideoEditFormComponent from "@/Components/Video/VideoEditFormComponent";
 import VideoFilesEditComponent from "@/Components/Video/VideoFilesEditComponent";
