+++ resources/ts/Pages/Forms/GroupRequest.tsx
@@ -1,10 +1,10 @@
 import CountryDropdown from "@/Components/Forms/CountryDropdown";
-import GroupLabel from "@/Components/Forms/GroupLabel";
-import InputLabel from "@/Components/Forms/InputLabel";
-import NumberInput from "@/Components/Forms/NumberInput";
-import SelectInput from "@/Components/Forms/SelectInput";
-import TextAreaInput from "@/Components/Forms/TextAreaInput";
-import TextInput from "@/Components/Forms/TextInput";
+import GroupLabel from "@/Elements/Forms/GroupLabel";
+import InputLabel from "@/Elements/Forms/InputLabel";
+import NumberInput from "@/Elements/Forms/NumberInput";
+import SelectInput from "@/Elements/Forms/SelectInput";
+import TextAreaInput from "@/Elements/Forms/TextAreaInput";
+import TextInput from "@/Elements/Forms/TextInput";
 import ToastBanner from "@/Components/Forms/ToastBanner";
 import PrimaryButton from "@/Elements/Buttons/PrimaryButton";
 import SecondaryButton from "@/Elements/Buttons/SecondaryButton";
