+++ resources/ts/Pages/LessonOrder/Create.tsx
@@ -1,8 +1,8 @@
 import ButtonLink from "@/Elements/Buttons/ButtonLink";
 import PrimaryButton from "@/Elements/Buttons/PrimaryButton";
-import InputLabel2 from "@/Components/Forms/InputLabel2";
-import NumberInput from "@/Components/Forms/NumberInput";
-import TextInput from "@/Components/Forms/TextInput";
+import InputLabel2 from "@/Elements/Forms/InputLabel2";
+import NumberInput from "@/Elements/Forms/NumberInput";
+import TextInput from "@/Elements/Forms/TextInput";
 import ToastBanner from "@/Components/Forms/ToastBanner";
 import ContentWrapper from "@/Layouts/ContentWrapper";
 import WrapperLayout from "@/Layouts/WrapperLayout";
