+++ resources/ts/Pages/LessonOrder/Edit.tsx
@@ -1,9 +1,9 @@
 import ButtonLink from "@/Elements/Buttons/ButtonLink";
 import PrimaryButton from "@/Elements/Buttons/PrimaryButton";
 import OrderInfoCard from "@/Components/Cards/OrderInfoCard";
-import InputLabel2 from "@/Components/Forms/InputLabel2";
-import NumberInput from "@/Components/Forms/NumberInput";
-import TextInput from "@/Components/Forms/TextInput";
+import InputLabel2 from "@/Elements/Forms/InputLabel2";
+import NumberInput from "@/Elements/Forms/NumberInput";
+import TextInput from "@/Elements/Forms/TextInput";
 import ToastBanner from "@/Components/Forms/ToastBanner";
 import Paragraph from "@/Components/Typography/Paragraph";
 import ParagraphContainer from "@/Components/Typography/ParagraphContainer";
